+++ prometheus-exporter/CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # Changelog
 
+## [1.2.0] - 2024-05-27
+
+### Added
+
+- Added cache for metrics endpoints
+- Added retrieval of metrics sequentially or in parallel
+
 ## [1.1.1] - 2024-04-02
 
 ### Fixed
@@ -21,6 +28,7 @@
 
 - First version of plugin with access to Thinger.io Database
 
+[1.2.0]: `https://github.com/thinger-io/plugins/compare/prometheus-exporter/v1.1.1...prometheus-exporter/v1.2.0`
 [1.1.1]: `https://github.com/thinger-io/plugins/compare/prometheus-exporter/v1.1.0...prometheus-exporter/v1.1.1`
 [1.1.0]: `https://github.com/thinger-io/plugins/compare/prometheus-exporter/v1.0.0...prometheus-exporter/v1.1.0`
 [1.0.0]: https://github.com/thinger-io/plugins/tree/prometheus-exporter/v1.0.0